requirements.txt
flask
flask-limiter
flask-cors
torch
transformers
pillow
requests
clip
